--- int_tests/all_tests.py
@@ -1,4 +1,0 @@
-from int_tests import solution_integ_test, folder_integ_test
-
-folder_integ_test
-solution_integ_test
