+++ tests/gcp_tests/all_tests.py
@@ -1,0 +1,8 @@
+from tests import unit_tests
+from tests.gcp_tests import folder_integ_test, solution_integ_test
+
+folder_integ_test
+solution_integ_test
+
+
+unit_tests
